I started objects for the reactor and a controller/sensor
so the "model for the reactor behaviour goes in the  "reactor" class code page. The other object takes a reading from the reactor, and can use it
typing report, or report temperature will allow there object properties to be called. In future these properties will be taken by the main program.